+++ backend/chat/src/main/java/com/adia/chat/repository/GroupeConversationRepository.java
@@ -8,6 +8,6 @@ import com.adia.chat.entity.GroupeConversation;
 import java.util.Optional;
 
 @Repository
-public interface GroupeConversationRepository extends JpaRepository<GroupeConversation, Long> {
-    Optional<GroupeConversation> findByConversationId(Long conversationId);
+public interface GroupeConversationRepository extends JpaRepository<GroupeConversation, Integer> {
+    Optional<GroupeConversation> findByConversationId(Integer conversationId);
 }
